+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## 12/01/2020
+* Extra Time: 1hr 54min
+* Added button for adding new item to list
+* Create completed method for removing completed items from localStoage
+* If changed to not completed a completed item will be re-added to localStorage
+* Changed item id to be the length of the todo array and increase by 1 for each new item added
+## 12/01/2020
 * Time remaining: 7min 44 seconds
 * Added localStoragec to store tasks
 * Tasks get fetched fron localStorage